+++ KEY_TIM_LED/Core/Src/main.c
@@ -25,7 +25,7 @@
 
 /* Private includes ----------------------------------------------------------*/
 /* USER CODE BEGIN Includes */
-
+#include "bsp_led.h"
 /* USER CODE END Includes */
 
 /* Private typedef -----------------------------------------------------------*/
@@ -187,7 +187,10 @@ void HAL_TIM_PeriodElapsedCallback(TIM_HandleTypeDef *htim)
     HAL_IncTick();
   }
   /* USER CODE BEGIN Callback 1 */
-
+  if (htim->Instance == TIM2)          /* TIM1 分支是 CubeMX 生成的 HAL 时基，别动 */
+  {
+    led_blink_tick_handler();
+  }
   /* USER CODE END Callback 1 */
 }